% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-coordinates.R
\docType{data}
\name{coordinates}
\alias{coordinates}
\title{Sample coordinates data}
\format{
A data frame with 20 rows and 2 variables:
\describe{
\item{Latitude}{Latitude, in WGS84 projection.}
\item{Longitude}{Longitude, in WGS84 projection.}
}
}
\usage{
coordinates
}
\description{
A dataset containing 20 sets of latitudes and longitudes in WGS84 projection.
}
\keyword{datasets}
