% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-locations.R
\docType{data}
\name{data_locations}
\alias{data_locations}
\title{Sample locations data}
\format{
A data frame with 20 rows and 3 variables:
\describe{
\item{Country code}{Country code.}
\item{Country}{Full name of a country.}
\item{Location}{Specific location, e.g. town or city name.}
}
}
\usage{
data_locations
}
\value{
A data frame with 20 rows and 3 variables of sample locations data.
}
\description{
A dataset containing 20 location information.
The raw dataset (\code{data_locations.csv}) is also provided in the directory \code{testdata}.
}
\keyword{datasets}
