% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{rtry_bind}
\alias{rtry_bind}
\title{Bind TRY data}
\usage{
rtry_bind(..., showOverview = TRUE)
}
\arguments{
\item{...}{A sequence of data frame to be combined by rows}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the combined data}
}
\value{
A data table of the input data
}
\description{
This function takes a sequence of data imported by \code{rtry_import()} and combined them by rows.
}
