% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{rtry_exclude}
\alias{rtry_exclude}
\title{Exclude TRY data}
\usage{
rtry_exclude(input = "", ..., showOverview = TRUE)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{...}{Criteria for excluding the data}

\item{showOverview}{Default \code{TRUE} displays the dimension of data table after excluding}
}
\value{
A data table of the input data after excluding the data
}
\description{
This function excludes data from the input data based on the specified criteria and the corresponding \code{ObsDataID}.
}
\examples{
\dontrun{
rtry_exclude(TRYdata, TraitID == 3116 & UnitName != "mm2 mg-1")
}
}
