% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{rtry_export}
\alias{rtry_export}
\title{Export pre-processed TRY data}
\usage{
rtry_export(data = "", output = "", quote = TRUE, encoding = "UTF-8")
}
\arguments{
\item{data}{The data to be saved}

\item{output}{Output path}

\item{quote}{Default \code{TRUE} inserts double quotes around any character or factor columns}

\item{encoding}{File encoding. Default \code{"UTF-8"}}
}
\description{
This function exports the pre-processed data as a CSV file.
If the specified output directory does not exist, it will be created.
}
\examples{
\dontrun{
rtry_export(TRYdata, "./output/TRYdata_processed.csv")
}
}
