% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_keyword.R
\name{rtry_filter_keyword}
\alias{rtry_filter_keyword}
\title{Filter data using keywords}
\usage{
rtry_filter_keyword(
  input = "",
  attribute = NULL,
  ...,
  caseSensitive = TRUE,
  exactMatch = TRUE,
  showOverview = TRUE
)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{attribute}{Attribute (column name) for filtering}

\item{\dots}{Values (keywords) for filtering}

\item{caseSensitive}{Default \code{TRUE} performs case-sensitive filtering}

\item{exactMatch}{Default \code{TRUE} performs exact match filtering, overrides all conflicting arguments}

\item{showOverview}{Default \code{TRUE} displays the dimension of data table after filtering}
}
\value{
A data table of the input data after filtering
}
\description{
This function filters data from the input data based on the specified keyword(s)
and the corresponding \code{ObservationID}.
}
\note{
This function filters data based on the unique identifier \code{ObservationID}
listed in the TRY data, therefore, if the column \code{ObservationID} has been removed, this function
will not work.
}
\examples{
# Filter observations which has the keywords "juvenile" or "unknown"
# within the column OrigValueStr, while excluding the whole observation
# Set the argument caseSensitive to FALSE to ignore case sensitivity
data_filtered <- rtry_filter_keyword(TRYdata_15160,
                   OrigValueStr,
                   c("juvenile", "unknown"),
                   caseSensitive = FALSE,
                   exactMatch = FALSE)

# Expected output:
# dim:   1605 28

# Learn more applications of the filtering function via the vignette (Workflow for general
# data preprocessing using rtry): vignette("rtry-workflow-general")
}
\seealso{
\code{\link{rtry_filter}}
}
