% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{rtry_geocoding}
\alias{rtry_geocoding}
\title{Perform Geocoding}
\usage{
rtry_geocoding(address = NULL, email = NULL)
}
\arguments{
\item{address}{String of an address}

\item{email}{String of an email address}
}
\value{
A data frame that contains latitudes (lat) and longitudes (lon) in WGS84 projection
}
\description{
This function uses \href{https://wiki.openstreetmap.org/wiki/Nominatim}{Nominatim}, a search engine for OpenStreetMap (OSM) data, to
perform Geocoding, i.e. converting an address into coordinates (latitudes, longitudes).
The data provided by OSM is free to use for any purpose, including commercial use,
and is governed by the distribution license \href{https://wiki.osmfoundation.org/wiki/Licence}{ODbL}.
}
\examples{
\dontrun{
rtry_geocoding("Hans-Knoell-Strasse 10, 07745 Jena, Germany", email = email)
}
}
\seealso{
\code{\link{rtry_revgeocoding}}
}
