% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rtry_import}
\alias{rtry_import}
\title{Import TRY data}
\usage{
rtry_import(
  input = "",
  separator = "\\t",
  encoding = "Latin-1",
  quote = "",
  showOverview = TRUE
)
}
\arguments{
\item{input}{Path to the text file downloaded from TRY}

\item{separator}{Data separator. Default \code{"\t"} for the TRY data output}

\item{encoding}{File encoding. Default \code{"Latin-1"}}

\item{quote}{Default \code{""} reads the fields as is. If the fields start with a double quote, use \code{"\""} instead}

\item{showOverview}{Default \code{TRUE} displays the input path, dimension and column names of the input data}
}
\value{
A data table of the input data
}
\description{
This function, by default, imports the text file exported from the TRY database
as a data.table for further processing. It can also be used to import other data file,
such as .csv file with comma as separator.
}
\examples{
\dontrun{
rtry_import("./data/7956.txt")
rtry_import("./data/coordinates.csv", separator = ",", encoding = "UTF-8", quote = "\"")
}
}
