% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_col.R
\name{rtry_merge_col}
\alias{rtry_merge_col}
\title{Merge two data frames (left join)}
\usage{
rtry_merge_col(x = "", y = "", baseOn = ObservationID, showOverview = TRUE)
}
\arguments{
\item{x}{The left data frame, imported by \code{rtry_import()} or in data table format}

\item{y}{The right data frame, imported by \code{rtry_import()} or in data table format}

\item{baseOn}{Default \code{ObservationID}, the common column used for merging}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the data after merging}
}
\value{
A data table of the merged data
}
\description{
This function merges two data frames based on a specified common column (by default: \code{ObservationID})
and returns all records from the left data frame (\code{x}) together with the matched records
from the right data frame (\code{y}) for further processing. In order words, this functions performs
a left join on the two provided data frames.
}
\examples{
\dontrun{
rtry_merge_col(data1, data2)
}
}
