% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_row.R
\name{rtry_merge_row}
\alias{rtry_merge_row}
\title{Merge two data frames (outer join)}
\usage{
rtry_merge_row(x = "", y = "", baseOn = ObservationID, showOverview = TRUE)
}
\arguments{
\item{x}{The left data frame, imported by \code{rtry_import()} or in data table format}

\item{y}{The right data frame, imported by \code{rtry_import()} or in data table format}

\item{baseOn}{Default \code{ObservationID}, the common column used for merging}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the data after merging}
}
\value{
A data table of the merged data
}
\description{
This function merges two data frames based on the row names of the  specified common column
(by default: \code{ObservationID}) and returns all rows from both tables, join records from the left (\code{x})
which have matching keys in the right data frame (\code{y}) for further processing.
In order words, this functions performs an outer join on the two provided data frames, i.e. the join table will contain
all records from both the tables.
}
\examples{
\dontrun{
rtry_merge_row(TRYdata1, TRYdata2)
}
}
