% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeocoding.R
\name{rtry_revgeocoding}
\alias{rtry_revgeocoding}
\title{Perform Reverse Geocoding}
\usage{
rtry_revgeocoding(lat_lon = NULL, email = NULL)
}
\arguments{
\item{lat_lon}{A data frame consisting latitude and longitude in WGS84 projection}

\item{email}{String of an email address}
}
\value{
A data frame that contains address
}
\description{
This function uses \href{https://wiki.openstreetmap.org/wiki/Nominatim}{Nominatim}, a search engine for OpenStreetMap data, to
perform reverse Geocoding, i.e. converting coordinates (latitudes, longitudes) into an address.
The data provided by OSM is free to use for any purpose, including commercial use,
and is governed by the distribution license \href{https://wiki.osmfoundation.org/wiki/Licence}{ODbL}.
}
\examples{
\dontrun{
rtry_revgeocoding(data.frame(50.91012, 11.56674), email = email)
}
}
\seealso{
\code{\link{rtry_geocoding}}
}
