% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_col.R
\name{rtry_rm_col}
\alias{rtry_rm_col}
\title{Remove TRY columns}
\usage{
rtry_rm_col(input, ..., showOverview = TRUE)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{...}{Column names to be removed}

\item{showOverview}{Default \code{TRUE} displays the dimension of the selected columns}
}
\value{
A data table of the remaining columns of the input data
}
\description{
This function removes specified columns from the imported data for further processing.
}
\examples{
\dontrun{
rtry_rm_col(TRYdata, Reference, Comment)
}
}
\references{
\href{https://www.rdocumentation.org/packages/dplyr/versions/0.7.8/topics/select}{dplyr::select()}
}
\seealso{
\code{\link{rtry_select_col}}
}
