% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{rtry_rmduplicates}
\alias{rtry_rmduplicates}
\title{Remove duplicates in TRY data}
\usage{
rtry_rmduplicates(input = "", showOverview = TRUE)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{showOverview}{Default \code{TRUE} displays the the dimension of data table after removal}
}
\value{
A data table of the input data after removing the duplicates
}
\description{
This function identifies and removed the duplicates from the input data.
Once the function is called and executed, the number of duplicates removed will be displayed on the console as reference.
}
