% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_col.R
\name{rtry_select_col}
\alias{rtry_select_col}
\title{Select columns}
\usage{
rtry_select_col(input = "", ..., showOverview = TRUE)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{...}{Column names to be selected}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the selected columns}
}
\value{
A data table of the selected columns of the input data
}
\description{
This function selects specified columns from the imported data and saves them in a new data table.
}
\examples{
# Select certain columns from the provided sample data (TRYdata_15160)
data_selected <- rtry_select_col(TRYdata_15160,
                   ObsDataID, ObservationID, AccSpeciesID, AccSpeciesName,
                   ValueKindName, TraitID, TraitName, DataID, DataName,
                   OriglName, OrigValueStr, OrigUnitStr, StdValue, UnitName,
                   OrigObsDataID, ErrorRisk, Comment)

# Expected output:
# dim:   1782 17
# col:   ObsDataID ObservationID AccSpeciesID AccSpeciesName ValueKindName
         TraitID TraitName DataID DataName OriglName OrigValueStr OrigUnitStr
         StdValue UnitName OrigObsDataID ErrorRisk Comment
}
\references{
\href{https://www.rdocumentation.org/packages/dplyr/versions/0.7.8/topics/select}{dplyr::select()}
}
\seealso{
\code{\link{rtry_rm_col}}
}
