% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_wider.R
\name{rtry_trans_wider}
\alias{rtry_trans_wider}
\title{Transform TRY data from long to wide table}
\usage{
rtry_trans_wider(
  input = "",
  names_from = NULL,
  values_from = NULL,
  values_fn = NULL,
  showOverview = TRUE
)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} function or in data table format}

\item{names_from}{The column(s) from which the output column names to be obtained}

\item{values_from}{The column(s) from which the output values to be obtained}

\item{values_fn}{(Optional) Function to be applied to the output values}

\item{showOverview}{Default \code{TRUE} displays the dimension of the result data table}
}
\value{
The transformed wide table
}
\description{
This function transforms the original long table format of the TRY data into a wide table format.
}
\examples{
\dontrun{
rtry_trans_wider(TRYdata,
                 names_from = c(TraitID, TraitName, UnitName),
                 values_from = c(StdValue, ErrorRisk),
                 values_fn = list(StdValue = mean, ErrorRisk = mean))
}
}
\references{
\href{https://www.rdocumentation.org/packages/tidytable/versions/0.5.7/topics/pivot_wider}{tidyr::pivot_wider()}
}
