% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-TRYdata-14833.R
\docType{data}
\name{TRYdata_14833}
\alias{TRYdata_14833}
\title{Sample TRY data (Request 14833)}
\format{
A data frame with 6409 rows and 28 variables:
\describe{
\item{LastName}{Surname of data contributor}
\item{FirstName}{First name of data contributor}
\item{DatasetID}{Unique identifier of contributed dataset}
\item{Dataset}{Name of contributed dataset}
\item{SpeciesName}{Original name of species}
\item{AccSpeciesID}{Unique identifier of consolidated species name}
\item{AccSpeciesName}{Consolidated species name}
\item{ObservationID}{Unique identifier for each observation}
\item{ObsDataID}{Unique identifier for each record}
\item{TraitID}{Unique identifier for traits (only if the record is a trait)}
\item{TraitName}{Name of trait (only if the record is a trait)}
\item{DataID}{Unique identifier for each sub-trait or context information}
\item{DataName}{Name of sub-trait or context information}
\item{OriginalName}{Original name of sub-trait or context information}
\item{OrigValueStr}{Original value as text string}
\item{OrigUnitStr}{Original unit as text string}
\item{ValueKindName}{Value kind (single measurement, mean, median, etc.)}
\item{OrigUncertaintyStr}{Original uncertainty as text string}
\item{UncertaintyName}{Kind of uncertainty (standard deviation, standard error, etc.)}
\item{Replicates}{Count of replicates}
\item{StdValue}{Standardized value: available for standardized traits}
\item{StdUnit}{Standard unit: available for standardized traits}
\item{RelUncertaintyPercent}{Relative uncertainty in \%}
\item{OrigObsDataID}{Unique identifier for duplicate entries}
\item{ErrorRisk}{Indication for outliers: distance to mean in standard deviations}
\item{Reference}{Reference to be cited if trait record is used in analysis}
\item{Comment}{Explanation for the \code{OriginalName} in the contributed dataset}
\item{V28}{Empty, an artefact due to the different software (MySQL >> R)}
...
}
}
\usage{
TRYdata_14833
}
\description{
A dataset requested from the \href{https://www.try-db.org/TryWeb/Prop0.php}{TRY Database}.
The request ID of this dataset is 14833, which contains \code{TraitID}: 3115, 3116, 3117
and \code{AccSpeciesID}: 10773, 35846, 45737.
}
\keyword{datasets}
