% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_col.R
\name{rtry_bind_col}
\alias{rtry_bind_col}
\title{Bind TRY data by columns}
\usage{
rtry_bind_col(..., showOverview = TRUE)
}
\arguments{
\item{...}{A sequence of data frame to be combined by columns}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the combined data}
}
\value{
A data table of the input data
}
\description{
This function takes a sequence of data imported by \code{rtry_import()}
and combines them by columns. The data should share the same number of rows.
}
\examples{
\dontrun{
rtry_bind_col(TRYdata1, TRYdata2)
}
}
