% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_col.R
\name{rtry_bind_col}
\alias{rtry_bind_col}
\title{Bind data by columns}
\usage{
rtry_bind_col(..., showOverview = TRUE)
}
\arguments{
\item{\dots}{A sequence of data frame to be combined by columns}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the combined data}
}
\value{
A data table of the input data
}
\description{
This function takes a sequence of data imported by \code{rtry_import()}
and combines them by columns. The data should share the same number of rows.
}
\examples{
# Assuming user has selected different columns as separated data tables
# and later on would like to combine as one for further processing
# Note: the binding process simply put the data side-by-side
data1 <- rtry_select_col(TRYdata_15160,
           ObsDataID, ObservationID, AccSpeciesID, AccSpeciesName,
           ValueKindName, TraitID, TraitName, DataID, DataName,
           OrigObsDataID, ErrorRisk, Comment)

data2 <- rtry_select_col(TRYdata_15160,
           OriglName, OrigValueStr, OrigUnitStr, StdValue, UnitName)

data <- rtry_bind_col(data1, data2)

Expected output:
dim:   1782 12
col:   ObsDataID ObservationID AccSpeciesID AccSpeciesName
       ValueKindName TraitID TraitName DataID DataName
       OrigObsDataID ErrorRisk Comment

dim:   1782 5
col:   OriglName OrigValueStr OrigUnitStr StdValue UnitName

dim:   1782 17
col:   ObsDataID ObservationID AccSpeciesID AccSpeciesName
       ValueKindName TraitID TraitName DataID DataName
       OrigObsDataID ErrorRisk Comment OriglName OrigValueStr
       OrigUnitStr StdValue UnitName
}
