% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_col.R
\name{rtry_bind_col}
\alias{rtry_bind_col}
\title{Bind data by columns}
\usage{
rtry_bind_col(..., showOverview = TRUE)
}
\arguments{
\item{\dots}{A list of data frames or data tables to be combined by columns.}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the combined data.}
}
\value{
An object of the same type as the first input.
}
\description{
This function takes a list of data frames or data tables and combines them by columns.
The data have to have the same number and sequence of rows.
}
\note{
A common attribute is not necessary (difference to the function \code{\link{rtry_join_left}} and \code{\link{rtry_join_outer}}):
the binding process simply puts the data side-by-side.
}
\examples{
# Assuming a user has selected different columns as separated data tables
# and later on would like to combine them as one for further processing.
data1 <- rtry_select_col(TRYdata_15160,
           ObsDataID, ObservationID, AccSpeciesID, AccSpeciesName, ValueKindName,
           TraitID, TraitName, DataID, DataName, OrigObsDataID, ErrorRisk, Comment)

data2 <- rtry_select_col(TRYdata_15160,
           OriglName, OrigValueStr, OrigUnitStr, StdValue, UnitName)

data <- rtry_bind_col(data1, data2)

# Expected messages:
# dim:   1782 12
# col:   ObsDataID ObservationID AccSpeciesID AccSpeciesName ValueKindName TraitID
#        TraitName DataID DataName OrigObsDataID ErrorRisk Comment
#
# dim:   1782 5
# col:   OriglName OrigValueStr OrigUnitStr StdValue UnitName
#
# dim:   1782 17
# col:   ObsDataID ObservationID AccSpeciesID AccSpeciesName ValueKindName TraitID
#        TraitName DataID DataName OrigObsDataID ErrorRisk Comment OriglName
#        OrigValueStr OrigUnitStr StdValue UnitName
}
\references{
This function makes use of the \code{\link[dplyr]{bind_cols}} function within the \code{dplyr} package.
}
\seealso{
\code{\link{rtry_bind_row}}, \code{\link{rtry_join_left}}, \code{\link{rtry_join_outer}}
}
