% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_row.R
\name{rtry_bind_row}
\alias{rtry_bind_row}
\title{Bind data by rows}
\usage{
rtry_bind_row(..., showOverview = TRUE)
}
\arguments{
\item{...}{A sequence of data frame to be combined by rows}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the combined data}
}
\value{
A data table of the input data
}
\description{
This function takes a sequence of data imported by \code{rtry_import()}
and combines them by rows. The data should share the same number of columns
and column names.
}
\examples{
# Combine the two provided sample data (TRYdata_15160 and TRYdata_15160)
data <- rtry_bind_row(TRYdata_15160, TRYdata_15161)

# Expected output:
# dim:   6409 28
# col:   LastName FirstName DatasetID Dataset SpeciesName
#        AccSpeciesID AccSpeciesName ObservationID ObsDataID TraitID
#        TraitName DataID DataName OriglName OrigValueStr OrigUnitStr
#        ValueKindName OrigUncertaintyStr UncertaintyName Replicates
#        StdValue UnitName RelUncertaintyPercent OrigObsDataID ErrorRisk
#        Reference Comment V28
}
