% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{rtry_explore}
\alias{rtry_explore}
\title{Explore data}
\usage{
rtry_explore(input = "", ..., sortBy = "", showOverview = TRUE)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} function or in data table format}

\item{...}{Attribute names to group together}

\item{sortBy}{(Optional) Specify the attribute name used to re-order the rows}

\item{showOverview}{Default \code{TRUE} displays the dimension of the result data table}
}
\value{
A data table of the unique values grouped by the desired attribute(s)
}
\description{
This function takes the data table imported by \code{rtry_import()} and converts it into
a grouped data table based on the specified column names.
Note that the output data are grouped by the first attribute if not specified using the parameter \code{sortBy}.
To provide a first understanding of the data, an additional column is added to show the total count within each group.
}
\examples{
# Explore the unique values in the provided sample data (TRYdata_15160)
# based on the attributes AccSpeciesID, DataID, DataName, TraitID and TraitName,
# and sorted by DataName
data_explore <- rtry_explore(TRYdata_15160,
                  AccSpeciesID, DataID, DataName, TraitID, TraitName,
                  sortBy = DataName)

# Expected output:
# dim:   235 6
}
\references{
\href{https://www.rdocumentation.org/packages/dplyr/versions/0.7.1/topics/group_by}{dplyr::group_by()}
}
