% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{rtry_filter}
\alias{rtry_filter}
\title{Filter TRY data}
\usage{
rtry_filter(input = "", ..., showOverview = TRUE)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{...}{Criteria for filtering}

\item{showOverview}{Default \code{TRUE} displays the dimension of data table after filtering}
}
\value{
A data table of the input data after removing the duplicates
}
\description{
This function filters data from the input data based on the specified criteria
}
\examples{
\dontrun{
rtry_filter(TRYdata, OrigValueStr \%in\% c("juvenile", "Juvenile", "juvenile, 6 weeks"))
}
}
\seealso{
\code{\link{rtry_filter_keyword}}
}
