% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{rtry_filter}
\alias{rtry_filter}
\title{Filter data}
\usage{
rtry_filter(input = "", ..., baseOn = ObservationID, showOverview = TRUE)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{\dots}{Criteria for filtering}

\item{baseOn}{Default \code{ObservationID}, the parameter used for filtering}

\item{showOverview}{Default \code{TRUE} displays the dimension of data table after filtering}
}
\value{
A data table of the input data after filtering
}
\description{
This function filters data from the input data based on the specified criteria
and by default this filtering is performed based on the corresponding \code{ObservationID}.
}
\note{
This function by default filters data based on the unique identifier \code{ObservationID}
listed in the TRY data, therefore, if the column \code{ObservationID} has been removed, this function
might not work (unless another attribute is defined when calling the function).
}
\examples{
# Filter observations where the plant developmental status (DataID 413)
# is "juvenile" or "unknown" while excluding the whole observation
data_filter <- rtry_filter(TRYdata_15160,
                 (DataID \%in\% 413) & (OrigValueStr \%in\% c("juvenile", "unknown")),
                 baseOn = ObservationID)

# Expected output:
# dim:   1618 28

# Learn more applications of the filtering function via the vignette (Workflow for general
# data preprocessing using rtry): vignette("rtry-workflow-general")
}
\seealso{
\code{\link{rtry_filter_keyword}}
}
