% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{rtry_geocoding}
\alias{rtry_geocoding}
\title{Perform Geocoding}
\usage{
rtry_geocoding(address = NULL, email = NULL)
}
\arguments{
\item{address}{String of an address}

\item{email}{String of an email address}
}
\value{
A data frame that contains latitudes (lat) and longitudes (lon) in WGS84 projection
}
\description{
This function uses Nominatim, a search engine for OpenStreetMap data, to
perform Geocoding, i.e. converting an address into coordinates (latitudes, longitudes).
For details, please refer to: \url{https://wiki.openstreetmap.org/wiki/Nominatim}.
}
\examples{
\dontrun{
rtry_geocoding("Hans-Knoell-Strasse 10, 07745 Jena, Germany", email = email)
}
}
\seealso{
\code{\link{rtry_revgeocoding}}
}
