% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{rtry_import}
\alias{rtry_import}
\title{Import TRY data}
\usage{
rtry_import(
  input = "",
  separator = "\\t",
  encoding = "UTF-8",
  showOverview = TRUE
)
}
\arguments{
\item{input}{Path to the text file downloaded from TRY}

\item{separator}{Data separator. Default \code{"\t"} for the TRY data output}

\item{encoding}{File encoding. Default \code{"UTF-8"}}

\item{showOverview}{Default \code{TRUE} displays the input path, dimension and column names of the input data}
}
\value{
A data table of the input data
}
\description{
This function imports the text file downloaded from the TRY database (\url{https://www.try-db.org})
as a data.table for further processing.
}
