% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rtry_import}
\alias{rtry_import}
\title{Import TRY data}
\usage{
rtry_import(
  input = "",
  separator = "\\t",
  encoding = "Latin-1",
  quote = "",
  showOverview = TRUE
)
}
\arguments{
\item{input}{Path to the text file downloaded from TRY}

\item{separator}{Data separator. Default \code{"\t"} for the TRY data output}

\item{encoding}{File encoding. Default \code{"Latin-1"}}

\item{quote}{Default \code{""} reads the fields as is. If the fields start with a double quote, use \code{"\""} instead}

\item{showOverview}{Default \code{TRUE} displays the input path, dimension and column names of the input data}
}
\value{
A data table of the input data
}
\description{
This function imports the text file downloaded from the TRY database (\url{https://www.try-db.org})
as a data.table for further processing.
}
\examples{
\dontrun{
rtry_import("./data/7956.txt")
rtry_import("./data/8200.txt", showOverview = FALSE)
}
}
