% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rtry_import}
\alias{rtry_import}
\title{Import data}
\usage{
rtry_import(
  input,
  separator = "\\t",
  encoding = "Latin-1",
  quote = "",
  showOverview = TRUE
)
}
\arguments{
\item{input}{Path to the data file.}

\item{separator}{Default \code{"\t"} for the TRY data output. Data separator.}

\item{encoding}{Default \code{"Latin-1"}. File encoding.}

\item{quote}{Default \code{""} reads the fields as is.
If the fields in the data file are by a double quote, use \code{"\""} instead.}

\item{showOverview}{Default \code{TRUE} displays the input path, the dimension and
the column names of the imported data.}
}
\value{
A \code{data.table}.
}
\description{
This function imports a data file as a \code{data.table} for further processing.
The default arguments are set to import tabulartor delimited data files in text
format (\code{.txt}) exported from the TRY database. It can also be used to
import other file formats, such as \code{.csv} files with comma separated values.
}
\examples{
# Example 1: Import data exported from the TRY database
# Specify file path to the raw data provided within the rtry package
input_path <- system.file("testdata", "TRYdata_15160.txt", package = "rtry")

# For own data and Windows users the path might rather look similar to this:
# input_path <- "C:/Users/User/Desktop/TRYdata_15160.txt"

# Import data file using rtry_import
input <- rtry_import(input_path)

# Explicit notation:
# input <- rtry_import(input_path, separator = "\t", encoding = "Latin-1",
#            quote = "", showOverview = TRUE)

# Expected message:
# input: ~/R/R-4.0.3/library/rtry/testdata/TRYdata_15160.txt
# dim:   1782 28
# col:   LastName FirstName DatasetID Dataset SpeciesName AccSpeciesID AccSpeciesName
#        ObservationID ObsDataID TraitID TraitName DataID DataName OriglName
#        OrigValueStr OrigUnitStr ValueKindName OrigUncertaintyStr UncertaintyName
#        Replicates StdValue UnitName RelUncertaintyPercent OrigObsDataID ErrorRisk
#        Reference Comment V28

# Example 2: Import CSV file
# Specify file path to the raw data provided within the rtry package
input_path <- system.file("testdata", "locations.csv", package = "rtry")

# Import data file using rtry_import
input <- rtry_import(input_path, separator = ",", encoding = "UTF-8",
           quote = "\"", showOverview = TRUE)

# Expected message:
# input: ~/R/R-4.0.3/library/rtry/testdata/locations.csv
# dim:   20 3
# col:   Country code Country Location
}
\references{
This function makes use of the \code{\link[data.table]{fread}} function
within the \code{data.table} package.
}
