% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_col.R
\name{rtry_merge_col}
\alias{rtry_merge_col}
\title{Merge two data frames (left join)}
\usage{
rtry_merge_col(x = "", y = "", baseOn = ObservationID, showOverview = TRUE)
}
\arguments{
\item{x}{The left data frame, imported by \code{rtry_import()} or in data table format}

\item{y}{The right data frame, imported by \code{rtry_import()} or in data table format}

\item{baseOn}{Default \code{ObservationID}, the common column used for merging}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the data after merging}
}
\value{
A data table of the merged data
}
\description{
This function merges two data frames based on a specified common column (by default: \code{ObservationID})
and returns all records from the left data frame (\code{x}) together with the matched records
from the right data frame (\code{y}), while discards all the records in the right data frame that
does not exist in the left data frame. In order words, this functions performs a left join on
the two provided data frames.
}
\examples{
# Assume user has obtained a unique list of auxiliary data (e.g. Longitude and Latitude)
# using rtry_select_aux() and would like to merge the georeferenced data into one data
# table according to the identifier ObservationID and while the Longitude data exists
lon <- rtry_select_aux(TRYdata_15160, Longitude)
lat <- rtry_select_aux(TRYdata_15160, Latitude)

georef <- rtry_merge_col(lon, lat)

# Expected output:
# dim:   97 2
# col:   ObservationID Longitude
#
# dim:   98 2
# col:   ObservationID Latitude
#
# dim:   97 3
# col:   ObservationID Longitude Latitude
}
\seealso{
\code{\link{rtry_merge_row}}, \code{\link{rtry_bind_col}}
}
