% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{rtry_revgeocoding}
\alias{rtry_revgeocoding}
\title{Perform Reverse Geocoding}
\usage{
rtry_revgeocoding(lat_lon = NULL, email = NULL)
}
\arguments{
\item{lat_lon}{A data frame consisting latitude and longitude in WGS84 projection}

\item{email}{String of an email address}
}
\value{
A data frame that contains address
}
\description{
This function uses Nominatim, a search engine for OpenStreetMap data, to
perform reverse Geocoding, i.e. converting coordinates (latitudes, longitudes) into an address.
For details, please refer to: \url{https://wiki.openstreetmap.org/wiki/Nominatim}.
}
\examples{
\dontrun{
rtry_revgeocoding(data.frame(50.91012, 11.56674), email = email)
}
}
