% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_aux.R
\name{rtry_select_aux}
\alias{rtry_select_aux}
\title{Select TRY auxiliary data in wide table format}
\usage{
rtry_select_aux(input = "", name = "", showOverview = TRUE)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{name}{The name of auxiliary data (\code{DataName} in the TRY data) to be selected}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the data after selection}
}
\value{
A data table of the selected auxiliary data of the input data in a wide table format
}
\description{
This function selects specified auxiliary data together with the \code{ObservationID}
from the imported data and transform it into a wide table format for further processing.
}
\examples{
\dontrun{
rtry_select_aux(TRYdata, Latitude)
}
}
