% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_aux.R
\name{rtry_select_aux}
\alias{rtry_select_aux}
\title{Select auxiliary data in wide table format}
\usage{
rtry_select_aux(input, name, showOverview = TRUE)
}
\arguments{
\item{input}{Input data frame or data table.}

\item{name}{The name of auxiliary data (\code{DataName} in the TRY data) to be selected.}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the selected data.}
}
\value{
An object of the same type as the input data.
}
\description{
This function selects one specified auxiliary data together with the \code{ObservationID}
from the imported data and transforms it into a wide table format for further processing.
It works for only one auxiliary data in \code{DataName} at a time.
}
\examples{
# Obtain a list of ObservationID and the corresponding auxiliary data of interest
# (e.g. "Latitude") from the provided sample data (TRYdata_15160)
Latitude <- rtry_select_aux(TRYdata_15160, Latitude)

# Expected message:
# dim:   98 2
# col:   ObservationID Latitude
}
\references{
This function makes use of the \code{\link[base]{subset}} and \code{\link[dplyr]{distinct}} functions
within the \code{base} and \code{dplyr} packages respectively. It also uses the functions
\code{\link{rtry_select_col}} and \code{\link{rtry_rm_col}}.
}
