% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_col.R
\name{rtry_select_col}
\alias{rtry_select_col}
\title{Select TRY columns}
\usage{
rtry_select_col(input = "", ..., showOverview = TRUE)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{...}{Column names to be selected}

\item{showOverview}{Default \code{TRUE} displays the dimension and column names of the selected columns}
}
\value{
A data table of the selected columns of the input data
}
\description{
This function selects specified columns from the imported data and saves them in a new data table.
}
\examples{
\dontrun{
rtry_select_col(data, DataID, DataName)
}
}
\references{
\href{https://www.rdocumentation.org/packages/dplyr/versions/0.7.8/topics/select}{dplyr::select()}
}
\seealso{
\code{\link{rtry_rm_col}}
}
