% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{rtry_select_col}
\alias{rtry_select_col}
\title{Select TRY columns}
\usage{
rtry_select_col(input = "", ..., showOverview = TRUE)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{...}{Column names to be selected}

\item{showOverview}{Default \code{TRUE} displays the dimension of the selected columns}
}
\value{
A data table of the selected columns of the input data
}
\description{
This function selects specified columns from the imported data for further processing.
}
\references{
\href{https://www.rdocumentation.org/packages/dplyr/versions/0.7.8/topics/select}{dplyr::select()}
}
