% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_row.R
\name{rtry_select_row}
\alias{rtry_select_row}
\title{Select rows}
\usage{
rtry_select_row(
  input = "",
  ...,
  getAuxiliary = FALSE,
  rmDuplicates = FALSE,
  showOverview = TRUE
)
}
\arguments{
\item{input}{Input data, imported by \code{rtry_import()} or in data table format}

\item{\dots}{Criteria for row selection}

\item{getAuxiliary}{Default \code{FALSE}, set to \code{TRUE} selects all auxiliary data based on the row selection criteria}

\item{rmDuplicates}{Default \code{FALSE}, set to \code{TRUE} calls the \code{rtry_rm_dup()} function}

\item{showOverview}{Default \code{TRUE} displays the dimension of the data after row selection}
}
\value{
A data table of the selected rows of the input data
}
\description{
This function selects rows based on specified criteria
and the corresponding \code{ObservationID} from the imported data for further processing.
}
\note{
This function by default filters data based on the unique identifier \code{ObservationID}
listed in the TRY data, therefore, if the column \code{ObservationID} has been removed, this function
will not work.
}
\examples{
# Select all traits records where DataID equals to 59 or 60
# together with all the corresponding auxiliary data without any duplicates
# within the provided sample data (TRYdata_15160)
data_selected <- rtry_select_row(TRYdata_15160,
                   (TraitID > 0) | (DataID \%in\% c(59, 60)),
                   getAuxiliary = TRUE,
                   rmDuplicates = TRUE)

# Expected output:
# 45 duplicates removed.
# dim:   1737 28
}
\seealso{
\code{\link{rtry_rm_dup}}
}
